fixing vercel error
attempting to fix vercel build error
